refactor(cli): thin side.ts — extract side-session orchestration to core (S2, #367)
Move orchestration algorithm and pure naming helpers out of cli/commands/side.ts
into packages/core/src/side-session.ts. CLI-edge deps (runtime construction,
agent command building, sendFirstTurnWhenReady) injected as closures. side.ts
drops from 363 to 219 LOC (-40%). Zero behavior change; 1555/1555 tests pass.
