+++ packages/core/src/index.ts
@@ -26,3 +26,4 @@ export * from "./crew-routing.js";
 export * from "./restart-daemon.js";
 export * from "./group-dispatch.js";
 export * from "./launch-workspace.js";
+export * from "./side-session.js";
